Test: search APIs test code
